Fix single page grid layout overflowing
Fix single page grid layout overflowing by using minmax(0, 1fr) instead
of 1fr. This works because 1fr means minmax(auto, 1fr) and when content
overflows auto is larger than the column should be. Setting the minimum
to 0 means 1fr is the maximum width, not auto which can be larger than
1fr depending on the content.